correcting placing of helper methods (one duplicate)
