+++ app/helpers/groups_helper.rb
@@ -7,22 +7,6 @@
 
 module GroupsHelper
 
-  def export_events_ical_button
-    type = params[:type].presence || 'Event'
-    if can?(:"export_#{type.underscore.pluralize}", @group)
-      action_button(I18n.t('event.lists.courses.ical_export_button'),
-        params.merge(format: :ics), :calendar)
-    end
-  end
-
-  def export_events_button
-    type = params[:type].presence || 'Event'
-    if can?(:"export_#{type.underscore.pluralize}", @group)
-      action_button(I18n.t('event.lists.courses.csv_export_button'),
-                    params.merge(format: :csv), :download)
-    end
-  end
-
   def tab_person_add_request_label(group)
     label = t('activerecord.models.person/add_request.other')
     count = Person::AddRequest.for_layer(group).count
